+++ backends/cuda/lib/cube/include/utl.h
@@ -4,8 +4,7 @@
 
 
 namespace CUBE_UTL_COPY {
-    cudaError_t copyDataFromHostToDevice(int Nx, int Ny, int Nz,complex_t* dest, complex_t* src, cudaStream_t stream = 0);
-    cudaError_t copyDataFromDeviceToHost(int Nx, int Ny, int Nz, complex_t* dest, complex_t* src, cudaStream_t stream = 0);
-    cudaError_t copyDataFromDeviceToDevice(int Nx, int Ny, int Nz, complex_t* dest, complex_t* src, cudaStream_t stream = 0);
+    cudaError_t copyDataFromHostToDevice(size_t Nx, size_t Ny, size_t Nz,complex_t* dest, complex_t* src, cudaStream_t stream = 0);
+    cudaError_t copyDataFromDeviceToHost(size_t Nx, size_t Ny, size_t Nz, complex_t* dest, complex_t* src, cudaStream_t stream = 0);
+    cudaError_t copyDataFromDeviceToDevice(size_t Nx, size_t Ny, size_t Nz, complex_t* dest, complex_t* src, cudaStream_t stream = 0);
 }
-
